+++ packages/core/src/main/cryptography/CryptographyService.ts
@@ -17,6 +17,7 @@
  *
  */
 
+const logdown = require('logdown');
 import APIClient = require('@wireapp/api-client');
 import {CRUDEngine} from '@wireapp/store-engine/dist/commonjs/engine/index';
 import {Cryptobox, store as CryptoboxStore} from '@wireapp/cryptobox';
@@ -27,7 +28,6 @@ import * as ProteusKeys from '@wireapp/proteus/dist/keys/root';
 import {PreKey as SerializedPreKey} from '@wireapp/api-client/dist/commonjs/auth/index';
 import CryptographyDatabaseRepository from './CryptographyDatabaseRepository';
 import {SessionPayloadBundle} from '../cryptography/root';
-import {ClientService} from '../client/root';
 import {OTRRecipients} from '@wireapp/api-client/dist/commonjs/conversation/index';
 
 export interface MetaClient extends RegisteredClient {
@@ -38,10 +38,15 @@ export interface MetaClient extends RegisteredClient {
 }
 
 export default class CryptographyService {
+  private logger: any = logdown('@wireapp/core/Account', {
+    logger: console,
+    markdown: false,
+  });
+
   public cryptobox: Cryptobox;
   private database: CryptographyDatabaseRepository;
 
-  constructor(private apiClient: APIClient, private storeEngine: CRUDEngine, private clientService: ClientService) {
+  constructor(private apiClient: APIClient, private storeEngine: CRUDEngine) {
     this.cryptobox = new Cryptobox(this.storeEngine);
     this.database = new CryptographyDatabaseRepository(this.storeEngine);
   }
@@ -51,6 +56,7 @@ export default class CryptographyService {
   }
 
   public async createCryptobox(): Promise<Array<SerializedPreKey>> {
+    this.logger.info('createCryptobox');
     const initialPreKeys: Array<ProteusKeys.PreKey> = await this.cryptobox.create();
 
     return initialPreKeys
@@ -65,6 +71,7 @@ export default class CryptographyService {
   }
 
   public decrypt(sessionId: string, encodedCiphertext: string): Promise<Uint8Array> {
+    this.logger.info('decrypt');
     const messageBytes: Uint8Array = Decoder.fromBase64(encodedCiphertext).asBytes;
     return this.cryptobox.decrypt(sessionId, messageBytes.buffer);
   }
@@ -74,6 +81,7 @@ export default class CryptographyService {
   }
 
   public async encrypt(plainText: Uint8Array, preKeyBundles: UserPreKeyBundleMap): Promise<OTRRecipients> {
+    this.logger.info('encrypt');
     const recipients: OTRRecipients = {};
     const encryptions: Array<Promise<SessionPayloadBundle>> = [];
 
@@ -107,6 +115,7 @@ export default class CryptographyService {
     plainText: Uint8Array,
     base64EncodedPreKey: string
   ): Promise<SessionPayloadBundle> {
+    this.logger.info('encryptPayloadForSession');
     let encryptedPayload;
 
     try {
@@ -124,12 +133,13 @@ export default class CryptographyService {
     return {sessionId, encryptedPayload};
   }
 
-  public async loadLocalClient(): Promise<RegisteredClient> {
+  public async initCryptobox(): Promise<void> {
+    this.logger.info('initCryptobox');
     const initialPreKeys: Array<ProteusKeys.PreKey> = await this.cryptobox.load();
-    return this.clientService.getLocalClient();
   }
 
   public deleteCryptographyStores(): Promise<boolean[]> {
+    this.logger.info('deleteCryptographyStores');
     return this.database.deleteStores();
   }
 }