+++ packages/api-client/CHANGELOG.md
@@ -3,6 +3,168 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.28"></a>
+## [0.6.28](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.27...@wireapp/api-client@0.6.28) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.27"></a>
+## [0.6.27](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.26...@wireapp/api-client@0.6.27) (2018-04-04)
+
+
+### Bug Fixes
+
+* **api-client:** Pass ClientType to Client.init after registration ([#455](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/issues/455)) ([1ef62ff](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/commit/1ef62ff))
+
+
+
+
+<a name="0.6.26"></a>
+## [0.6.26](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.25...@wireapp/api-client@0.6.26) (2018-04-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.25"></a>
+## [0.6.25](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.24...@wireapp/api-client@0.6.25) (2018-04-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.24"></a>
+## [0.6.24](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.23...@wireapp/api-client@0.6.24) (2018-04-03)
+
+
+### Bug Fixes
+
+* **api-client,store-engine:** Always create new Dexie instance on Engine.init() ([#448](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/issues/448)) ([27623ee](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/commit/27623ee))
+
+
+
+
+<a name="0.6.23"></a>
+## [0.6.23](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.22...@wireapp/api-client@0.6.23) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.22"></a>
+## [0.6.22](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.21...@wireapp/api-client@0.6.22) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.21"></a>
+## [0.6.21](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.20...@wireapp/api-client@0.6.21) (2018-04-03)
+
+
+### Bug Fixes
+
+* **api-client,cryptobox:** Show log messages during tests ([#435](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/issues/435)) ([559759d](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/commit/559759d))
+
+
+
+
+<a name="0.6.20"></a>
+## [0.6.20](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.19...@wireapp/api-client@0.6.20) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.19"></a>
+## [0.6.19](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.18...@wireapp/api-client@0.6.19) (2018-03-30)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.18"></a>
+## [0.6.18](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.17...@wireapp/api-client@0.6.18) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.17"></a>
+## [0.6.17](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.16...@wireapp/api-client@0.6.17) (2018-03-29)
+
+
+### Bug Fixes
+
+* **api-client,core:** Login by phone number removed from request ([#434](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/issues/434)) ([c5123fd](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/commit/c5123fd))
+
+
+
+
+<a name="0.6.16"></a>
+## [0.6.16](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.15...@wireapp/api-client@0.6.16) (2018-03-29)
+
+
+### Bug Fixes
+
+* **api-client:** Use updateOrCreate when updating access token ([#433](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/issues/433)) ([10dbae5](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/commit/10dbae5))
+
+
+
+
+<a name="0.6.15"></a>
+## [0.6.15](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.14...@wireapp/api-client@0.6.15) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.14"></a>
+## [0.6.14](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.13...@wireapp/api-client@0.6.14) (2018-03-28)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.13"></a>
+## [0.6.13](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.12...@wireapp/api-client@0.6.13) (2018-03-27)
+
+
+### Bug Fixes
+
+* **api-client,core:** Optional client type parameter for init ([#422](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/issues/422)) ([29750e1](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/commit/29750e1))
+
+
+
+
+<a name="0.6.12"></a>
+## [0.6.12](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.11...@wireapp/api-client@0.6.12) (2018-03-27)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
+<a name="0.6.11"></a>
+## [0.6.11](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.10...@wireapp/api-client@0.6.11) (2018-03-26)
+
+
+
+
+**Note:** Version bump only for package @wireapp/api-client
+
 <a name="0.6.10"></a>
 ## [0.6.10](https://github.com/wireapp/wire-web-packages/tree/master/packages/api-client/compare/@wireapp/api-client@0.6.9...@wireapp/api-client@0.6.10) (2018-03-26)