+++ packages/cbor/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.14"></a>
+## [3.0.14](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.13...@wireapp/cbor@3.0.14) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cbor
+
+<a name="3.0.13"></a>
+## [3.0.13](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.12...@wireapp/cbor@3.0.13) (2018-04-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cbor
+
+<a name="3.0.12"></a>
+## [3.0.12](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.11...@wireapp/cbor@3.0.12) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @wireapp/cbor
+
 <a name="3.0.11"></a>
 ## [3.0.11](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.10...@wireapp/cbor@3.0.11) (2018-03-26)