+++ packages/core/CHANGELOG.md
@@ -3,6 +3,212 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.43"></a>
+## [2.1.43](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.42...@wireapp/core@2.1.43) (2018-04-06)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.42"></a>
+## [2.1.42](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.41...@wireapp/core@2.1.42) (2018-04-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.41"></a>
+## [2.1.41](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.40...@wireapp/core@2.1.41) (2018-04-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.40"></a>
+## [2.1.40](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.39...@wireapp/core@2.1.40) (2018-04-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.39"></a>
+## [2.1.39](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.38...@wireapp/core@2.1.39) (2018-04-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.38"></a>
+## [2.1.38](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.37...@wireapp/core@2.1.38) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.37"></a>
+## [2.1.37](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.36...@wireapp/core@2.1.37) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.36"></a>
+## [2.1.36](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.35...@wireapp/core@2.1.36) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.35"></a>
+## [2.1.35](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.34...@wireapp/core@2.1.35) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.34"></a>
+## [2.1.34](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.33...@wireapp/core@2.1.34) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.33"></a>
+## [2.1.33](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.32...@wireapp/core@2.1.33) (2018-04-03)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.32"></a>
+## [2.1.32](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.31...@wireapp/core@2.1.32) (2018-03-30)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.31"></a>
+## [2.1.31](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.30...@wireapp/core@2.1.31) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.30"></a>
+## [2.1.30](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.29...@wireapp/core@2.1.30) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.29"></a>
+## [2.1.29](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.28...@wireapp/core@2.1.29) (2018-03-29)
+
+
+### Bug Fixes
+
+* **api-client,core:** Login by phone number removed from request ([#434](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/issues/434)) ([c5123fd](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/commit/c5123fd))
+
+
+
+
+<a name="2.1.28"></a>
+## [2.1.28](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.27...@wireapp/core@2.1.28) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.27"></a>
+## [2.1.27](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.26...@wireapp/core@2.1.27) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.26"></a>
+## [2.1.26](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.25...@wireapp/core@2.1.26) (2018-03-28)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.25"></a>
+## [2.1.25](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.24...@wireapp/core@2.1.25) (2018-03-27)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.24"></a>
+## [2.1.24](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.23...@wireapp/core@2.1.24) (2018-03-27)
+
+
+### Bug Fixes
+
+* **api-client,core:** Optional client type parameter for init ([#422](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/issues/422)) ([29750e1](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/commit/29750e1))
+
+
+
+
+<a name="2.1.23"></a>
+## [2.1.23](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.22...@wireapp/core@2.1.23) (2018-03-27)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.22"></a>
+## [2.1.22](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.21...@wireapp/core@2.1.22) (2018-03-27)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.21"></a>
+## [2.1.21](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.20...@wireapp/core@2.1.21) (2018-03-27)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.20"></a>
+## [2.1.20](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.19...@wireapp/core@2.1.20) (2018-03-26)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
+<a name="2.1.19"></a>
+## [2.1.19](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.18...@wireapp/core@2.1.19) (2018-03-26)
+
+
+
+
+**Note:** Version bump only for package @wireapp/core
+
 <a name="2.1.18"></a>
 ## [2.1.18](https://github.com/wireapp/wire-web-packages/tree/master/packages/core/compare/@wireapp/core@2.1.17...@wireapp/core@2.1.18) (2018-03-26)